Added Creation Warnings

+++ lib/budget_files/budget.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 
 class Budget {
   String goal;
-  int budgetAmount;
+  double budgetAmount;
   DateTime? startDate;
   DateTime? endDate;
   IconData icon;
-  int totalUsed = 0;
+  double totalUsed = 0;
   Color color = Colors.blue;
 
   Budget(this.goal, this.budgetAmount, this.startDate, this.endDate, this.icon,
